Fix sort order for merging search results
If two or more IPs have the same last_part_at time then sort the
ips by id which is unique.  This ensures that we get the
callsign details records in the same order as the IPs and
we don't miss results when we build the resultant table
for display.