Display numeric bzid instead of blank if one is available for an IP
If the bzid is missing for a player join line but another
join for the same callsign has a bzid on the same IP then
the results are better with a single line of detail which
encompasses the entire date range for both joins.

This merges detail lines for a bug that leaves out bzid when
a player is ghosted.