Only redirect to the login page for access errors if we are not already logged in
There's no point prompting for login credentials if the user is already logged in
and they don't have access to the function they are trying to reach.  Normally users
can't navigate to restricted functions but there's nothing stopping them from
bookmarking restricted pages while logged in or entering the URL directly in the
location bar of the browser.