Skip log messages that have already been loaded
We can now reprocess old logs and have it pick up the time
of the last log in the database.  The parser skips all of
the logs until that timestamp is reached and checks each
log with that timestamp for duplicate records - skipping the duplicates

This allows us to reprocess old logs and skip the ones
already loaded.  Now we don't have to be careful reprocessing
old data.