Setup to skip log messages already processed
Read the last log message created for this server and save
the time in @last_log_time.  We'll use this to skip logs
in the stream with a timestamp earlier than this log.

This allows for rerunning old logs in the parser without
creating duplicate records.  If you need to force reprocessing
of old logs delete the log records first.