git_version: Return git version information for the project
Set project version number to reflect working directory state.

Releases (without git) will include a version number file.  If this
exists we just use whatever version number it has.  Otherwise we
if we're in a git repository use the details of git describe
and include '-dirty' if any tracked files are modified but not committed.

Signed-off-by: Bernt Hansen <bernt@norang.ca>