Tweak log search to gain speed
Remove the 'logged_at desc' from the order clause as it
was redundant and made the generated query unable to use
any of the existing indexes. Log_messages.id are created
in the same order as the messages are in the raw logs
which technically should be in the same order as 'logged_at'.
The id field is indexed which is where the speed boost
comes from. Removal increased performance of the log
search from ~9000ms to ~550ms from click to render.

 Signed-off-by: Joe Van Overberghe <joevano@gmail.com>