Replace the raw sql query for searching by callsign
Utilize the new Callsign.callsign_like method to create a matching
list of callsigns. Walk through the association in Callsigns to Ips and
collect the id field for all of the matching ips and add them to an array.
Pass the array of ids into an 'Ip.find' so that all three methods (callsign,
IP, and Hostname) collect the exact same object.

Signed-off-by: Joe Van Overberghe <joevano@gmail.com>