Modify Ips model to allow for searching by 'callsign like'.
Modify the Ips model to allow searching for :callsign_like.
Went with a raw sql query because the rails model generated
queries were just plain slow. Add the ability to specify a
limit to the number of Ips returned with the default being
500 if none limit parameter is passed.

Modify the SearchController to use the new Ip.callsign_like().
This eliminated the need to find the callsigns, then find the
ips for those callsigns and then finally pull all of the Ip model objects.

Signed-off-by: Joe Van Overberghe <joevano@gmail.com>