Add basic caching to the Callsign model
Add basic caching to the callsign model to reduce the number
of lookups being done on the table. Players come and generally
have quite a bit of activity so it makes sense to look them up
once and remember them for subsequent actions.

Add the ability to clear the cache to aid in testing. Expose
the length of the cache, for testing as well.

Signed-off-by: Joe Van Overberghe <joevano@gmail.com>