Create socket server and make log_parser multi-server aware
Create socket server to receive multiple streams of log file data
with records in the format of <hostname>|<port>|log_line, and
the corresponding unit tests.

Modify log_parser to track the last_log_times and dates
from multiple servers simultaneously.

Change log_parser tests to take into account the multi-server awareness
of caching in the log_parser. Extended the log_parser class to allow
for clearing of the cache and querying the value of the @log_parser
hash while in the testing environment.

Signed-off-by: Joe Van Overberghe <joevano@gmail.com>